chore(디렉토리 구조 변경) : Global, Domain 형태로 구조 변경
