--- src/main/java/com/clover/habbittracker/config/JpaConfig.java
@@ -1,10 +1,0 @@
-package com.clover.habbittracker.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-
-}
