+++ src/main/java/com/clover/habbittracker/domain/habbit/api/HabbitController.java
@@ -1,0 +1,7 @@
+package com.clover.habbittracker.domain.habbit.api;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HabbitController {
+}
