+++ src/main/java/com/clover/habbittracker/domain/habbit/dto/HabbitRequest.java
@@ -1,0 +1,8 @@
+package com.clover.habbittracker.domain.habbit.dto;
+
+import lombok.Getter;
+
+@Getter
+public class HabbitRequest {
+	String content;
+}
