+++ src/main/java/com/clover/habbittracker/domain/habbit/entity/Habbit.java
@@ -1,0 +1,37 @@
+package com.clover.habbittracker.domain.habbit.entity;
+
+
+import com.clover.habbittracker.domain.member.entity.BaseEntity;
+import com.clover.habbittracker.domain.member.entity.Member;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Entity
+@Table(name = "habbit")
+public class Habbit extends BaseEntity {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	private String content;
+
+	@ManyToOne
+	@JoinColumn(name = "memberId")
+	private Member member;
+
+
+}
