+++ src/main/java/com/clover/habbittracker/domain/habbit/repository/HabbitRepository.java
@@ -1,0 +1,9 @@
+package com.clover.habbittracker.domain.habbit.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.clover.habbittracker.domain.habbit.entity.Habbit;
+
+public interface HabbitRepository extends JpaRepository<Habbit,Long> {
+
+}
