+++ src/main/java/com/clover/habbittracker/domain/member/dto/MemberRequest.java
@@ -1,0 +1,11 @@
+package com.clover.habbittracker.domain.member.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class MemberRequest {
+	String nickName;
+	String profileImgUrl;
+}
