+++ src/main/java/com/clover/habbittracker/global/config/JpaConfig.java
@@ -1,0 +1,10 @@
+package com.clover.habbittracker.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+
+}
