+++ src/main/java/com/clover/habbittracker/global/jwt/JwtException.java
@@ -1,0 +1,8 @@
+package com.clover.habbittracker.global.jwt;
+
+public class JwtException extends RuntimeException{
+
+	public JwtException(String message) {
+		super(message);
+	}
+}
