+++ src/main/java/com/clover/habbittracker/global/oauth/dto/SocialUser.java
@@ -1,0 +1,15 @@
+package com.clover.habbittracker.global.oauth.dto;
+
+
+
+public interface SocialUser {
+	String getEmail();
+
+	String getNickName();
+
+	String getProvider();
+
+	String getOauthId();
+
+	String getProfileImgUrl();
+}
