--- src/main/java/com/clover/habbittracker/habbit/api/HabbitController.java
@@ -1,7 +1,0 @@
-package com.clover.habbittracker.habbit.api;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HabbitController {
-}
