--- src/main/java/com/clover/habbittracker/habbit/dto/HabbitRequest.java
@@ -1,8 +1,0 @@
-package com.clover.habbittracker.habbit.dto;
-
-import lombok.Getter;
-
-@Getter
-public class HabbitRequest {
-	String content;
-}
