--- src/main/java/com/clover/habbittracker/habbit/service/HabbitService.java
@@ -1,9 +1,0 @@
-package com.clover.habbittracker.habbit.service;
-
-import com.clover.habbittracker.habbit.dto.HabbitRequest;
-import com.clover.habbittracker.habbit.dto.HabbitRespose;
-
-public interface HabbitService {
-	HabbitRespose register(Long MemberId, HabbitRequest request);
-
-}
