--- src/main/java/com/clover/habbittracker/jwt/JwtException.java
@@ -1,8 +1,0 @@
-package com.clover.habbittracker.jwt;
-
-public class JwtException extends RuntimeException{
-
-	public JwtException(String message) {
-		super(message);
-	}
-}
