--- src/main/java/com/clover/habbittracker/member/dto/MemberRequest.java
@@ -1,11 +1,0 @@
-package com.clover.habbittracker.member.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class MemberRequest {
-	String nickName;
-	String profileImgUrl;
-}
