--- src/main/java/com/clover/habbittracker/oauth/dto/SocialUser.java
@@ -1,15 +1,0 @@
-package com.clover.habbittracker.oauth.dto;
-
-
-
-public interface SocialUser {
-	String getEmail();
-
-	String getNickName();
-
-	String getProvider();
-
-	String getOauthId();
-
-	String getProfileImgUrl();
-}
